Ignore git version and configuration differences.
From the `git status` docs:
```
--porcelain
Give the output in an easy-to-parse format for scripts.
This is similar to the short output, but will remain stable across Git versions
and regardless of user configuration. See below for details.
```

For example, `isRepoClean()` returns `false` even when the repo is clean when using
the `status.branch` config option:

```
$ git config status.branch true
$ git status
## master...origin/master
$ gitsem minor
2017/05/25 18:16:48 repo isn't clean
$ git config status.branch false
$ git status # prints nothing
$ gitsem minor # works
```